Updated the four operators
- Added comments in all operators.
- Added progress logging during operator execution
- Cleaned up variable names

+++ plugins/operators/load_dimension.py
@@ -11,23 +11,33 @@ class LoadDimensionOperator(BaseOperator):
     def __init__(self,
                  redshift_conn_id="",
                  table = "",
-                 stl_statement = "",
-                 #missing target table
+                 sql_statement = "",
+                 truncate_insert = True, # default value is true
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id=redshift_conn_id
-        self.stl_statement = stl_statement
+        self.sql_statement = sql_statement
         self.table=table
+        self.truncate_insert=truncate_insert
 
     def execute(self, context):
-        self.log.info('LoadDimensionOperator not implemented yet')
-        #dimension must be emptied before a load
-
+        """
+        This operator loads data from the staging tables to the dimension tables. It requires the following parameters:
+            redshift_conn_id : contains the connection details to the data warehouse in Amazon Redshift (from in Airflow)
+            table : the name of the destination dimension table
+            sql_statement : the SQL query to get the data to be inserted in the destination table
+            truncate-insert : boolean indicating whether the dimension table is to be emptied before inserting the data
+        """
+        self.log.info("Loading data into the {} table".format(self.table))
+        #connect to redshift
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
+        #if the parameter is true, we empty the dimension before loading the data
+        if self.truncate_insert:
+            self.log.info("Deleting rows in the {} table".format(self.table))
+            redshift.run("DELETE FROM {}".format(self.table))
 
-        self.log.info("DELETE FROM {}".format(self.table))
-        redshift.run("DELETE FROM {}".format(self.table))
-        completeSql = "INSERT INTO {} ".format(self.table) + self.stl_statement
-        self.log.info("Inserting to dim table")
+        completeSql = "INSERT INTO {} ".format(self.table) + self.sql_statement
+        self.log.info(completeSql)
         redshift.run(completeSql)
+        self.log.info("Insertion into the {} table completed.".format(self.table))
